+++ config/log.go
@@ -10,6 +10,7 @@ import (
 	"net/url"
 	"time"
 
+	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp"
 	"go.opentelemetry.io/otel/exporters/stdout/stdoutlog"
 	"go.opentelemetry.io/otel/log"
@@ -61,7 +62,7 @@ func logProcessor(ctx context.Context, processor LogRecordProcessor) (sdklog.Pro
 		}
 		return sdklog.NewSimpleProcessor(exp), nil
 	}
-	return nil, fmt.Errorf("unsupported log processor type, must be one of simple or batch")
+	return nil, errors.New("unsupported log processor type, must be one of simple or batch")
 }
 
 func logExporter(ctx context.Context, exporter LogRecordExporter) (sdklog.Exporter, error) {
@@ -79,6 +80,8 @@ func logExporter(ctx context.Context, exporter LogRecordExporter) (sdklog.Export
 		switch *exporter.OTLP.Protocol {
 		case protocolProtobufHTTP:
 			return otlpHTTPLogExporter(ctx, exporter.OTLP)
+		case protocolProtobufGRPC:
+			return otlpGRPCLogExporter(ctx, exporter.OTLP)
 		default:
 			return nil, fmt.Errorf("unsupported protocol %q", *exporter.OTLP.Protocol)
 		}
@@ -152,4 +155,46 @@ func otlpHTTPLogExporter(ctx context.Context, otlpConfig *OTLP) (sdklog.Exporter
 	}
 
 	return otlploghttp.New(ctx, opts...)
+}
+
+func otlpGRPCLogExporter(ctx context.Context, otlpConfig *OTLP) (sdklog.Exporter, error) {
+	var opts []otlploggrpc.Option
+
+	if otlpConfig.Endpoint != nil {
+		u, err := url.ParseRequestURI(*otlpConfig.Endpoint)
+		if err != nil {
+			return nil, err
+		}
+		// ParseRequestURI leaves the Host field empty when no
+		// scheme is specified (i.e. localhost:4317). This check is
+		// here to support the case where a user may not specify a
+		// scheme. The code does its best effort here by using
+		// otlpConfig.Endpoint as-is in that case
+		if u.Host != "" {
+			opts = append(opts, otlploggrpc.WithEndpoint(u.Host))
+		} else {
+			opts = append(opts, otlploggrpc.WithEndpoint(*otlpConfig.Endpoint))
+		}
+		if u.Scheme == "http" {
+			opts = append(opts, otlploggrpc.WithInsecure())
+		}
+	}
+	if otlpConfig.Compression != nil {
+		switch *otlpConfig.Compression {
+		case compressionGzip:
+			opts = append(opts, otlploggrpc.WithCompressor(*otlpConfig.Compression))
+		case compressionNone:
+			// none requires no options
+		default:
+			return nil, fmt.Errorf("unsupported compression %q", *otlpConfig.Compression)
+		}
+	}
+	if otlpConfig.Timeout != nil && *otlpConfig.Timeout > 0 {
+		opts = append(opts, otlploggrpc.WithTimeout(time.Millisecond*time.Duration(*otlpConfig.Timeout)))
+	}
+	if len(otlpConfig.Headers) > 0 {
+		opts = append(opts, otlploggrpc.WithHeaders(toStringMap(otlpConfig.Headers)))
+	}
+
+	return otlploggrpc.New(ctx, opts...)
 }